update README.md with import supported by node
It describes how to import with the es modules, but they are not supported by current versions of nodeJS. It's [experimental](https://nodejs.org/api/esm.html). You can use them with babel (or another transpilator), but it is not included in the project and will add a layer of complexity. I updated the readme with the currently supported native syntax.