Fix navbar rendering issues with theme support (#1014)
* Fix navbar rendering issues with theme support

This corrects rendering the navbar using convention used by Bootswatch
and BS4, which usually either defaults to default styles for navbar or
uses combination of the navbar custom style and navbar custom baground
color (bg-*). IF `bg-*` is not present, this could lead to the
transparent background color for navbar when expanded.

The change contains a backward compatiblity fix for default theme, which
was done by not overriding `bg-*` value in navbars, but by using custom
background color as it was used before some changes in BS4. So by
default theme will use backward compatible override for background
color, not one that comes with the BS4 for 'bg-light' class.