Migrate from googleapis npm package to @googleapis/drive and sheets
The googleapis’s package is combination of @googleapis/drive and @googleapis/sheets and other google APIs.  It is 130MB and adds about 40 sec to the startup time because of its massive size.  Devs can add various @googleapis/* packages based on their need.  As of this commit @googleapis/drive and sheets are 2 MB and 646kB respectively.
Ref: https://github.com/sahat/hackathon-starter/issues/1234