fix: replace express-flash with our own middleware (#1461)
Both express-flash and connect-flash are unmaintained (last updated around 12 years ago). connect-flash has known bugs, and we’ve been patching it with patch-package for months. Continuing to rely on these packages introduces technical debt and fragility. This change deprecates the use of these two unmaintained packages in favor of a short middleware with the same functionality without external dependencies.

---------

Co-authored-by: Yashar Fakhari <6448697+YasharF@users.noreply.github.com>